Removing tests
Two of these tests work (the 'entirely_replaced' tests) but are of
questionable value.

The other two tests (the 'contains_alternate_file' tests) are nice, but
are really just convoluted tests of the AppConf functionality. Also,
they don't work, because when they attempt to override the setting they
are testing against it is too late, as AppConf has already consumed the
original setting.